# backend/openapi.yaml
openapi: 3.1.0
info:
  title: BudgetBuddy API
  version: "1"
  description: |
    Operational header behavior: responses include `X-Request-Id`.
    If request provides `X-Request-Id`, the API echoes it; otherwise it generates one.
servers:
- url: /api
security:
- BearerAuth: []
paths:
  /auth/register:
    post:
      security: []
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Username already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /auth/login:
    post:
      security: []
      summary: Login a user
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /auth/refresh:
    post:
      security: []
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /auth/logout:
    post:
      summary: Logout (revoke refresh token)
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  headers:
    X-Request-Id:
      description: Correlation id for tracing requests. Propagated when provided by client, generated otherwise.
      schema:
        type: string
    Retry-After:
      description: Seconds to wait before retrying after rate limiting.
      schema:
        type: integer
        minimum: 1
...
  schemas:
    ProblemDetails:
      type: object
      required: [ type, title, status ]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
    User:
      type: object
      required: [ id, username, currency_code ]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          pattern: '^[a-zA-Z0-9_]{3,20}$'
        currency_code:
          type: string
          enum: [ USD, COP, EUR, MXN ]
    RegisterRequest:
      type: object
      required: [ username, password, currency_code ]
      properties:
        username:
          type: string
          pattern: '^[a-zA-Z0-9_]{3,20}$'
        password:
          type: string
          minLength: 12
        currency_code:
          type: string
          enum: [ USD, COP, EUR, MXN ]
    AuthResponse:
      type: object
      required: [ user, access_token, refresh_token, access_token_expires_in ]
      properties:
        user:
          $ref: "#/components/schemas/User"
        access_token:
          type: string
        refresh_token:
          type: string
        access_token_expires_in:
          type: integer
          description: Seconds until access token expiration
    LoginRequest:
      type: object
      required: [ username, password ]
      properties:
        username:
          type: string
          pattern: '^[a-zA-Z0-9_]{3,20}$'
        password:
          type: string
          minLength: 12
    RefreshRequest:
      type: object
      required: [ refresh_token ]
      properties:
        refresh_token:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT